script: Move Element attribute getters/setters into attributes.rs
Following #43714, move the attribute getter/setter methods from
element.rs into the separate attributes.rs file. This reduces the
size of element.rs and groups attribute-related methods together.

The following methods were moved:
- set_atomic_attribute
- set_bool_attribute
- get_url_attribute / set_url_attribute
- get_trusted_type_url_attribute
- get_trusted_html_attribute
- get_string_attribute / set_string_attribute
- get_nullable_string_attribute / set_nullable_string_attribute
- get_tokenlist_attribute / set_tokenlist_attribute
- set_atomic_tokenlist_attribute
- set_int_attribute
- get_uint_attribute / set_uint_attribute
